Missing checks when generating RSA keys.
Based on the following changes, but with
some differences:

 Add https://build.opensuse.org/projects/openSUSE:Factory/packages/mozilla-nss/files/nss-fips-rsa-keygen-strictness.patch

 HG changeset patch
 User M. Sirringhaus <msirringhaus@suse.de>
 Date 1584305670 -3600
      Sun Mar 15 21:54:30 2020 +0100
 Node ID 2f570c6952d8edfc1ad9061cd3830f202eec1960
 Parent  557f9009507c9e70941dbe39965028049e1ef5a2
 commit 4b8c0eac6b092717157b4141c82b4d76ccdc91b3
 Author: Hans Petter Jansson <hpj@cl.no>
    Patch 16: nss-fips-rsa-keygen-strictness.patch

Firstly, add in the test changes to mpprime.c in
all cases, not just FIPS mode.

Secondly, retain the "Check that the exponent is not smaller than 65537"
case in non-FIPS mode, as moving this into FIPS mode only breaks the
test gtests.sh: #8009: 'RSATest: expThreeTest'.

Signed-off-by: Jeremy Allison <jallison@ciq.com>